feat: multiple environment deployments
Leverages Mustache and separate contract JSON files to allow multiple
deployments based on configuration variables.

Here we have two initial configuations; one for our own playground
environment and another for production.
