+++ .prettierignore
@@ -1,0 +1,2 @@
+src/constants/contracts.ts.mustache
+subgraph.yaml.mustache
