+++ config/production.json
@@ -1,0 +1,25 @@
+{
+  "network": "matic",
+  "fuelActivatedBlock": 20637829,
+  "v1EndBlock": 20384301,
+  "BaseGETV1": {
+    "address": "0x308e44cA2153C61103b0DC67Fd038De650912b73",
+    "startBlock": 15980526
+  },
+  "EventMetadataStorageV1": {
+    "address": "0xcDA348fF8C175f305Ed8682003ec6F8743067f79",
+    "startBlock": 15980516
+  },
+  "BaseGETV2": {
+    "address": "0xbce1b23c7544422f1E2208d29A6A3AA9fAbAB250",
+    "startBlock": 20363107
+  },
+  "EventMetadataStorageV2": {
+    "address": "0x08C2aF3F01A36AD9F274ccE77f6f77cf9aa1dfC9",
+    "startBlock": 20363134
+  },
+  "EconomicsGETV2": {
+    "address": "0x07faA643ad0eE4ee358d5E101573A5fdfBEcD0a9",
+    "startBlock": 20637233
+  }
+}
