+++ package.json
@@ -3,11 +3,13 @@
   "license": "MIT",
   "scripts": {
     "codegen": "graph codegen",
+    "create:local": "graph create --node http://localhost:8020/ get-protocol-subgraph",
+    "prepare:playground": "mustache config/playground.json subgraph.yaml.mustache > subgraph.yaml && mustache config/playground.json src/constants/contracts.ts.mustache > src/constants/contracts.ts",
+    "prepare:production": "mustache config/production.json subgraph.yaml.mustache > subgraph.yaml && mustache config/production.json src/constants/contracts.ts.mustache > src/constants/contracts.ts",
+    "deploy:local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 get-protocol-subgraph",
+    "deploy:playground": "graph deploy --product hosted-service getprotocol/playground-get-protocol-subgraph",
+    "deploy:production": "graph deploy --product hosted-service getprotocol/get-protocol-subgraph",
     "build": "graph build",
-    "deploy": "graph deploy --node https://api.studio.thegraph.com/deploy/ get-protocol-subgraph",
-    "create-local": "graph create --node http://localhost:8020/ get-protocol-subgraph",
-    "remove-local": "graph remove --node http://localhost:8020/ get-protocol-subgraph",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 get-protocol-subgraph",
     "prepare": "husky install",
     "check": "eslint . && prettier --check --ignore-path .gitignore .",
     "lint": "eslint .",
@@ -26,6 +28,7 @@
     "eslint-config-prettier": "^8.3.0",
     "husky": ">=7.0.4",
     "lint-staged": ">=11.2.3",
+    "mustache": "^4.2.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.4.1",
     "typescript": "^4.4.4"
