+++ src/mappings/v2/eventMetadataStorage.ts
@@ -3,7 +3,7 @@ import {
   EventMetadataStorageV2 as EventMetadataStorageContract,
   NewEventRegistered,
 } from "../../../generated/EventMetadataStorageV2/EventMetadataStorageV2";
-import { BIG_INT_ZERO, EVENT_METADATA_STORAGE_ADDRESS_V2 } from "../../constants";
+import { BIG_DECIMAL_ZERO, BIG_INT_ZERO, EVENT_METADATA_STORAGE_ADDRESS_V2 } from "../../constants";
 import { getEvent, createUsageEvent } from "../../entities";
 
 export function handleNewEventRegistered(e: NewEventRegistered): void {
@@ -33,5 +33,5 @@ export function handleNewEventRegistered(e: NewEventRegistered): void {
 
   event.save();
 
-  createUsageEvent(e, event, BIG_INT_ZERO, "NEW_EVENT", e.params.orderTime, e.params.getUsed);
+  createUsageEvent(e, event, BIG_INT_ZERO, "NEW_EVENT", e.params.orderTime, BIG_DECIMAL_ZERO);
 }
